Only add base and segment items for clocks
Removes the dividers and border. Now a similar visual is achieved by shrinking
the segments a bit, leaving room between each other and the outer border. This
means there is some space that you can grab if you want to select or move the
clock. The segments are a bit rounded, but not perfectly circular.