Remove base
Changes how a clock is drawn. Now the segments themselves have borders. The
base is replaced by a pin, a small circle that can be selected if you want to
delete or copy a clock. Especially if the label is empty.

All items making up a clock are connected to each other in sequence: second
item to first, third to second etc. This means you can move by clicking on
segments and dragging.