Use player colour for all components
Instead of light or dark grey the lines and pin now only have the players
colour. This should mean that the clocks are visible on any background where
other things like the pointer are. The segments are now semitransparent to
still make it easy to see how many there are when filled.